requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-debug-toolbar==2.1
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.4
pyTelegramBotAPI==3.6.6
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
